fix chain name in asset to sell list not changing
